chore(OccurrenceStatuses): update model, factory, serializer, migration
