+++ database/factories/OccurrenceStatusFactory.php
@@ -3,16 +3,16 @@
 declare(strict_types=1);
 
 use Faker\Generator as Faker;
-use VOSTPT\Models\ProCivOccurrenceStatus;
+use VOSTPT\Models\OccurrenceStatus;
 
 /*
 |--------------------------------------------------------------------------
-| ProCivOccurrenceStatus Factories
+| OccurrenceStatus Factories
 |--------------------------------------------------------------------------
 |
 */
 
-$factory->define(ProCivOccurrenceStatus::class, function (Faker $faker) {
+$factory->define(OccurrenceStatus::class, function (Faker $faker) {
     return [
         'code' => $faker->unique()->numberBetween(1, 255),
         'name' => $faker->unique()->sentence(),
